# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DermaAI - Skin Cancer Detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses (history)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new skin analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /analyses/upload:
    post:
      operationId: uploadImage
      tags: [analyses]
      summary: Upload an image for analysis (returns analysis result)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
  /analyses/{id}/report:
    get:
      operationId: downloadReport
      tags: [analyses]
      summary: Download PDF report for an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF report info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportInfo"
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List doctors
      parameters:
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AI Doctor Assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
components:
  schemas:
...
    Analysis:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        patientName:
          type: string
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        condition:
          type: string
        conditionDetails:
          type: ["string", "null"]
        prediction:
          type: string
        riskLevel:
          type: string
        confidenceScore:
          type: number
        cancerType:
          type: ["string", "null"]
        abcdeScore:
          type: ["string", "null"]
        lesionArea:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        segmentedImageUrl:
          type: ["string", "null"]
        heatmapImageUrl:
          type: ["string", "null"]
        riskFactors:
          $ref: "#/components/schemas/RiskFactors"
        riskProgressData:
          type: array
          items:
            $ref: "#/components/schemas/RiskDataPoint"
        explainableAiReasons:
          type: array
          items:
            type: string
        reportId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - patientName
        - prediction
        - riskLevel
        - confidenceScore
        - createdAt
    AnalysisInput:
      type: object
      properties:
        patientName:
          type: string
        age:
          type: integer
        gender:
          type: string
        conditionDetails:
          type: string
        imageData:
          type: string
        riskFactors:
          $ref: "#/components/schemas/RiskFactors"
      required:
        - patientName
    ImageUploadInput:
      type: object
      properties:
        patientName:
          type: string
        conditionDetails:
          type: string
        imageData:
          type: string
        age:
          type: integer
        gender:
          type: string
        riskFactors:
          $ref: "#/components/schemas/RiskFactors"
      required:
        - patientName
        - imageData
    RiskFactors:
      type: object
      properties:
        sunExposure:
          type: ["string", "null"]
        skinType:
          type: ["string", "null"]
        familyHistory:
          type: ["boolean", "null"]
        age:
          type: ["integer", "null"]
        immuneCondition:
          type: ["boolean", "null"]
      required: []
    RiskDataPoint:
      type: object
      properties:
        month:
          type: string
        riskScore:
          type: number
      required:
        - month
        - riskScore
    AnalysisResult:
      type: object
      properties:
        analysisId:
          type: integer
        prediction:
          type: string
        riskLevel:
          type: string
        confidenceScore:
          type: number
        cancerType:
          type: string
        abcdeScore:
          type: string
        lesionArea:
          type: string
        recommendations:
          type: string
        imageUrl:
          type: string
        segmentedImageUrl:
          type: string
        heatmapImageUrl:
          type: string
        riskProgressData:
          type: array
          items:
            $ref: "#/components/schemas/RiskDataPoint"
        explainableAiReasons:
          type: array
          items:
            type: string
        similarCases:
          type: array
          items:
            $ref: "#/components/schemas/SimilarCase"
        reportId:
          type: string
        isHighRisk:
          type: boolean
      required:
        - analysisId
        - prediction
        - riskLevel
        - confidenceScore
        - recommendations
        - isHighRisk
    SimilarCase:
      type: object
      properties:
        id:
          type: integer
        patientName:
          type: string
        prediction:
          type: string
        riskLevel:
          type: string
        confidenceScore:
          type: number
        outcome:
          type: string
        imageUrl:
          type: ["string", "null"]
        date:
          type: string
      required:
        - id
        - patientName
        - prediction
        - riskLevel
        - confidenceScore
        - date
    ReportInfo:
      type: object
      properties:
        reportId:
          type: string
        downloadUrl:
          type: string
        qrCodeData:
          type: string
        generatedAt:
          type: string
      required:
        - reportId
        - downloadUrl
        - qrCodeData
        - generatedAt
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        hospital:
          type: string
        city:
          type: string
        country:
          type: string
        experience:
          type: integer
        rating:
          type: number
        phone:
          type: string
        email:
          type: string
        isOnline:
          type: boolean
        avatar:
          type: ["string", "null"]
      required:
        - id
        - name
        - specialty
        - hospital
        - city
        - country
        - experience
        - rating
        - isOnline
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
        analysisId:
          type: ["integer", "null"]
      required:
        - message
    ChatResponse:
      type: object
      properties:
        id:
          type: integer
        response:
          type: string
        timestamp:
          type: string
      required:
        - id
        - response
        - timestamp
...