# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  hospital: string;
  city: string;
  country: string;
  experience: number;
  rating: number;
  phone?: string;
  email?: string;
  isOnline: boolean;
  /** @nullable */
  avatar?: string | null;
}